back/pyproject.toml
[project]
name = "Sonari"
version = "2025.01.1"
description = "Audio Annotation Tool"
authors = [{ name = "Santiago Martinez", email = "santiago.mbal@gmail.com" }, { name = "Artur Sterz", email = "sterz@trackit.systems" }]
dependencies = [
    "uvicorn[standard]>=0.20.0",
    "aiosqlite>=0.18.0",
    "passlib>=1.7.4",
    "fastapi>=0.101.0",
    "pydantic-settings>=2.0.2",
    "fastapi-users[sqlalchemy]>=12.1.1",
    "cachetools>=5.3.1",
    "asyncache>=0.3.1",
    "fastapi-pagination>=0.12.8",
    "alembic>=1.13.0",
    "bcrypt==4.0.1",
    "colorama>=0.4.6",
    "openpyxl>=3.1.5",
    "astral>=3.2",
]
requires-python = ">=3.11,<3.12"
readme = "README.md"
keywords = ["audio", "annotation", "tool", "bioacoustics", "machine learning"]
license = { file = "LICENSE" }

[project.urls]
Homepage = "https://github.com/trackIT-Systems/sonari"
Documentation = "https://github.com/trackIT-Systems/sonari"
"Isue Tracker" = "https://github.com/trackIT-Systems/sonari/issues"

[project.optional-dependencies]
docs = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.4.14",
    "mkdocstrings[python]>=0.24.0",
]
postgre = ["asyncpg>=0.29.0", "psycopg2-binary>=2.9.9"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
artifacts = ["src/sonari/migrations/versions/", "src/sonari/statics/"]

[tool.pyright]
venvPath = "."
venv = ".venv"
include = ["src"]

[tool.pydocstyle]
convention = "numpy"

[tool.ruff]
line-length = 79
target-version = "py311"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 60

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "Q", "I", "D"]
ignore = ["D1"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.isort]
known-local-folder = ["src", "sonari"]

[tool.rye]
dev-dependencies = [
    "pytest>=7.2.2",
    "coverage>=7.2.1",
    "black>=23.1.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.23.3",
    "tox>=4.7.0",
    "flake8>=6.1.0",
    "pyright>=1.1.363",
    "ruff>=0.1.7",
    "icecream>=2.1.3",
    "pytest-xdist>=3.5.0",
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.4.14",
    "mkdocstrings[python]>=0.24.0",
    "pytest-cov>=4.1.0",
    "crowsetta==4.0.0.post2",
    "cython==3.0.10",
    "email-validator==2.1.1",
    "importlib-resources==5.13.0",
    "matplotlib==3.7.5",
    "pydantic==2.7.1",
    "rasterio==1.3.10",
    "scikit-learn==1.3.2",
    "scipy==1.10.1",
    "shapely==2.0.4",
    "soundfile==0.12.1",
    "xarray==2023.1",
    "soundevent==2.0.0",
]
